# modules/reservation/reservation-headless/reservation-headless-impl/rest-openapi.yaml
components:
  schemas:
...
    Participant:
      description: Contains information about a booking participant.
      properties:
        participantId:
          description: The primary key for the Participant entity.
          type: integer
          format: int64
          readOnly: true
        userId:
          description: Set to 0 if not associated with any existing Liferay user. Will use existing liferay user's attributes and ignore other fields if userID is valid.
          type: integer
          format: int64
        fullName:
          description: Participant's full name.
          type: string
        companyName:
          description: Participant's company name.
          type: string
        emailAddress:
          description: Participant's email address.
          type: string
      required: 
        - userId
      type: object
...
openapi: 3.0.1
info:
  description: "API for accessing Reservation details."
  license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
  title: "Headless Reservation"
  version: v1.0
paths:
...
  "/participants":
    get:
      operationId: getParticipantsPage
      tags: ["Participant"]
      description: Retrieve the list of offices. Results can be paginated, filtered, searched and sorted.
      parameters:
        - in: query
          name: filter
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: pageSize
          schema:
            type: integer
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: sort
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Participant"
                type: array
            application/xml:
              schema:
                items:
                  $ref: "#/components/schemas/Participant"
                type: array
    post:
      operationId: postParticipant
      tags: ["Participant"]
      description: Create a new participant entity.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Participant"
          application/xml:
            schema:
              $ref: "#/components/schemas/Participant"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Participant"
            application/xml:
              schema:
                $ref: "#/components/schemas/Participant"
  "/participants/{participantId}":
    get:
      operationId: getParticipant
      tags: ["Participant"]
      description: Retrieves the participant via its ID.
      parameters:
        - name: participantId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Participant"
            application/xml:
              schema:
                $ref: "#/components/schemas/Participant"
    put:
      operationId: putParticipant
      tags: ["Participant"]
      description: Updates and overwrites the target participant.
      parameters: 
        - name: participantId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Participant"
          application/xml:
            schema:
              $ref: "#/components/schemas/Participant"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Participant"
            application/xml:
              schema:
                $ref: "#/components/schemas/Participant"
    delete:
      operationId: deleteParticipant
      tags: ["Participant"]
      description: Deletes the target participant
      parameters: 
        - name: participantId
          in: path
          required: true
          schema:
            type: string
      responses:
        204:
          description: ""
          content:
            application/json: {} 
            application/xml: {} 
...